@@ -35,8 +35,8 @@
   },
   "peerDependencies": {
     "livekit-client": "^0.18.1",
-    "react": "^15.0.0 || ^16.0.0 || ^17.0.0",
-    "react-dom": "^15.0.0 || ^16.0.0 || ^17.0.0"
+    "react": ">=15",
+    "react-dom": ">=15"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-common-types": "^6.1.1",
@@ -45,7 +45,7 @@
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^25.1.4",
     "@types/node": "^12.12.38",
-    "@types/react": "^17.0.38",
+    "@types/react": "^18.0.9",
     "@types/react-dom": "^17.0.13",
     "@typescript-eslint/eslint-plugin": "^3.10.1",
     "@typescript-eslint/parser": "^3.10.1",
@@ -64,10 +64,10 @@
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "^4.0.3",
-    "typescript": "~4.2.3"
+    "typescript": "^4.2.3"
   },
   "files": [
     "dist",
