@@ -1,7 +1,7 @@
 {
   "name": "livekit-react-example",
   "homepage": ".",
-  "version": "0.8.3",
+  "version": "1.0.0-dev1",
   "private": true,
   "scripts": {
     "start": "react-scripts start",
@@ -14,9 +14,9 @@
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@livekit/react-components": "^0.8.3",
     "livekit-client": "^0.18.1",
-    "react": "^15.0.0 || ^16.0.0 || ^17.0.0",
+    "react": "^18.1.0",
     "react-aspect-ratio": "^1.0.50",
-    "react-dom": "^15.0.0 || ^16.0.0 || ^17.0.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
@@ -26,7 +26,7 @@
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^25.1.4",
     "@types/node": "^12.12.38",
-    "@types/react": "^17.0.38",
+    "@types/react": "^18.0.9",
     "@types/react-dom": "^17.0.13",
     "@types/react-router-dom": "5.3.3",
     "eslint": "^8.13.0",
@@ -42,7 +42,7 @@
     "gh-pages": "^2.2.0",
     "prettier": "^2.6.2",
     "react-scripts": "^4.0.3",
-    "typescript": "~4.2.3"
+    "typescript": "^4.2.3"
   },
   "eslintConfig": {
     "extends": "react-app"
