@@ -1,5 +1,6 @@
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom'
 import App from './App'
 import './index.css'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const root = createRoot(document.getElementById('root') as Element)
+root.render(<App />)
